Add Othello player that can play with external programs using GTP (#316)
GTP (Go Text Protocol) is a text protocol that allows
Go (the game, not the programming language) programs to
play with each other. Due to the similarity in the board
representation the protocol is used by Othello/Reversi programs.

Add Othello player that uses the protocol to play
with external programs. That makes it easy to test how well
the AI performs against other AI bots.

Add some additional methods to the Player interface - startGame,
endGame and notify. Those methods are essential to keep the
state synced with external program.

Update Arena to use those methods if present to keep it
backward compatible and not to force other implementations
to implement class. To this end GTPOthelloPlayer is callable
so that players can still be a simple function.